Refactor and enhance filter validation and processing
Improved biquad filter configuration validation by adding frequency checks for standard filter types.

Simplified biquad filter parameter updates by consolidating coefficient computation logic.

Enhanced convolution filter error handling for coefficient loading.

Added checks for finite delay values in delay filter configuration.

Normalized coefficient scaling in differential equation filters.

Streamlined gain filter processing to use scalar multiplication for muted states.

Updated lookahead limiter to handle parameter changes more gracefully.

Adjusted volume filter initialization to sync with target volume at cold start.

Improved mixer processing to handle input channel count mismatches.

Refined pipeline computation to simplify mixer retrieval logic.

Enhanced public configuration handling to prevent samplerate overrides for specific device types.

Added checks for NaN and invalid delta values in fader volume adjustments.

Increased public update interval from 100ms to 1000ms for better performance.

Improved signal range retrieval in processing engine.

Enhanced channel label retrieval to handle missing labels more robustly.

Updated global peak retrieval and reset functions to check for engine capabilities.

Removed unused async poly resampler profile handling and streamlined configuration validation.

Improved audio resampler validation for cutoff frequency.

Added channel count checks in slip resampler processing.

Removed unnecessary optimizations in synchronous resampler.

Updated lock-free ring buffer to include atomic float max update functionality.

Enhanced raw and wav reader implementations for better error handling and data integrity.

Adjusted wav writer header logic to ensure proper fact chunk handling based on seekability.

Updated engine_processing_loop_set_pipeline to accept a flag for transferring filter states during pipeline swaps.

Modified pipeline_transfer_state to conditionally transfer filter and processor states based on the new flag.

Simplified configuration diffing by removing the config_change_t structure and related functions, directly returning change types.

Enhanced tests to ensure proper handling of pipeline state and configuration changes.

Improved documentation to reflect changes in state transfer logic and configuration handling.

sync to upstream 99e9724 from b438410.

BTW, In recent upstream commits, we found more and more enhancements coming from our existing code bases, including but not limited to moving hot reload to control thread, stricter param validation, etc.

+++ src/Audio/spectrum_analyzer.h
@@ -91,7 +91,7 @@ size_t spectrum_analyzer_get_fft_n(const spectrum_analyzer_t* analyzer);
 spectrum_status_t spectrum_analyzer_compute(spectrum_analyzer_t* analyzer,
                                             audio_history_buffer_t* buffer,
                                             const size_t* channel,
-                                            float min_freq, float max_freq,
+                                            double min_freq, double max_freq,
                                             size_t n_bins, size_t samplerate,
                                             spectrum_result_t* out_result);
